add comments for network/connection classes
I've added some basic explanations for the connection classes. The basic
architecture is that each "network" class must manage the connections of the
appropriate type. It occurs to me that this may not be the best way to
architect things. This strategy will lead to connection management code that is
duplicated between the different network classes. One way to deal with that
would be for the different network classes to inherit from a common base that
has most of the actual connection management code.