fix multiple submit
Our "main.js" is compiled and minified from some coffeescript code that no
longer exists in the repo (see commits made when the gulpfile bower and whatnot
were still here). It will need to be completely replaced at some point, but for
now I'm just building in the compiled, minified code. There was a problem in
the old build process that kept concatenating javascript to the same file,
making a lot of the code duplicated. This had the weird effect of performing
the form submit over and over again, once for each time the code had been
built. Since I had built it something like 17 times, it would submit the form
17 times. I've now fixed that and the form is submitted once.

The email addresses are now saved to a .dat file in bin/. This will be much
easier to manage than the emails that formspree is sending me. Not to mention
formspree is only free up to 100 emails per month.