peers toJSON and asyncConnectManyfromJSON
You can now "save" the peers to a JSON object, and then import that same object
later on to re-establish connections using asyncConnectManyFromJSON. This will
be used in DBPeers - we will want to save the list of peers using .toJSON, and
when we retrieve the list of peers, we will want to try to re-establish
connections to them. For instance, suppose you close your browser window. When
you open it again, you will probably want to connect to the same peers you were
connected to before. Maybe not all of the will work, but some will. This will
eliminate an unnecessary peer discovery stage every single time you reload your
browser.

Note: I am currently not putting any thought into how to handle multiple
browser windows open at the same time. Presently, multiple windows will use the
same database but not the same network connections. It is possible, in
principle, to use a shared web worker to manage the network connections. That
is almost certainly what we're going to want to do to make sure we don't
establish multiple p2p connections to the same people. Using shared workers
will also make managing the database simpler - there is only one thread that
accesses the database.