DBPeers - for saving/retrieving peer info from db
The information about what peers we are connected to is stored in memory. It is
useful to be able to store this information to the database, so that when we
close the current process and open it again later, we can automatically
reconnect to the same peers as before. Some of those peers may have also
disconnected, but some may still be around. It's a good place to start without
having to do peer discovery.

The architecture of this will be as follows. CorePeers will depend on both
Peers and DBPeers. Peers is what actually creates and receives connections. It
will handle all the logic of this. However, CorePeers, having access both to
Peers and DBPeers, will make sure that the current state of Peers is saved to
the db.