5 proxy servers, one rendezvous server
To make it easier to test the (upcoming) p2p connections in the UI, we will
want to be able to run several different apps simultaneously that share the
same rendezvous server. I now have the rendezvous server running on port 3031.
The normal app server is running on port 3030. When you run gulp serve, it runs
5 instances of browserSync, which acts as a proxy. Those are running on ports
3040 - 3044. In other words, you can open up, say, localhost:3040 and
localhost:3041 to view the app, sharing the same rendezvous server running on
port 3031.

Breakdown of ports:
- 3030: Normal app server. Run the app at localhost:3030
- 3031: PeerJS rendezvous server. Maintains lists of peers and their IDs.
- 3040: A browserSync server. Uses rendezvous server on port 3031.
- 3041: A browserSync server. Uses rendezvous server on port 3031.
- 3042: A browserSync server. Uses rendezvous server on port 3031.
- 3043: A browserSync server. Uses rendezvous server on port 3031.
- 3044: A browserSync server. Uses rendezvous server on port 3031.

All of these ports are for development purposes, and public-facing ports have
not yet been decided.