CorePeers can now respond to pong messages
The big picture of this commit is that the basic structure of CorePeers is now
in place and it can listen to messages and respond. This is where almost all of
the messaging logic will go. For now, it just handles ping messages. A next
step will be to make it handle ContentAuth messages, i.e. where someone is
sending us content.

Details of some things that changed:
- Move handling ping message outside of Connection and into CorePeers
- Rename all handler methods to prefix with "handle" instead of "on"
- Make CorePeers an EventEmitter
- Add validator for ping message
- Move some Network handlers to distinct methods
- Add some handlers for Peers to watch connections/messages