asyncGetAllPeers
A development-only method for getting a list of peers on the rendezvous server.
This will be useful for easily connecting to a bunch of peers during the
development stage of Datt. However, as explained in the comments, this should
probably be removed before we launch to a general audience, because it doesn't
scale, and doesn't work with web socket peers. Not to mention it relies on the
rendezvous server to exist.