display # of peers in UI accurately on connection
Every time a new connection is established, the number of peers as displayed in
the UI is updated accurately to reflect the number of peers we are currently
connected to. There is a small problem with this as currently implemented,
which is that we are not tracking disconnects. So when a peer disconnects, the
number will no longer be accurate.