received contentauth now saved to db
You can now test in the browser that content is sent from browser to browser
and then actually saved in the db. To test this, open up localhost:3040 and
localhost:3041 at the same time. Post some new content in localhost:3040 - then
refresh localhost:3041 and you should see the new content on the front-page.