views -> react, lib -> core
It's slightly more explanatory that "views" be labeled "react", and same for
"lib" being labeled "core". It is the case that react/* is where all the
dattreact code is and core/* is where the dattcore code is. The directories now
match the names. Hypothetically we may now also add other front-ends with
different names. However, more important, I think there will some new files
which are NOT a part of "dattreact" or "dattcore", and therefore do not belong
mixed in with either of them. That is the case for spawn, which I will probably
be moving to a different directory soon.