move spawn to util/ directory
"spawn" really has nothing to do with dattcore in particular, and I would like
to be able to use it from dattreact without having to pull in any files from
the same directory as dattcore. That's because the *only* way we should be
using dattcore from within dattreact is via a global dattcore object - we
shouldn't be using "require" to get modules, which would lead to too much
duplication when things are built (since dattcore goes to datt-core.js and
dattreact goes to datt-react.js). For a small handful of general utilities that
can be "required" from either dattreact or dattcore, those will go in the new
util folder. Theoretically anything in the util folder could also simply be its
own distinct module. I won't worry about that for now.