remove blank files
These files were here to demonstrate what the architecture would probably be
like, although they are blank. They are either for "payments" or "actions". A
payment is a kind of action. They are stored in the database, hence db-payment
and db-action. We will add these back once we fill them in.