BIP 44 Account
This is an "account" in the sense of BIP44, which is a particular way of
deriving key addresses with BIP32. This class will be used by the upcoming
BIP44Wallet class which will let you manage multiple accounts.

This code probably properly belongs in fullnode and not in Datt, however
because it uses AsyncCrypto, it will be in Datt for now. fullnode has an
entirely differnent way of managing workers which I may ultimately change to be
more like Datt. In any case, for now this code goes here.