BIP44 Wallet outline
BIP44 specifies a standard way to create new addresses from a master extended
private key. This commit sketches out an outline for how to do this. We have a
master mnemonic which can be used to derive the master extended private key and
master extended public key (masterxprv and masterxpub). Then we have a list of
accounts which contain addresses which can be derived from the masterxprv.