rename: AsyncCrypto -> CryptoWorkers
Every method in Datt that returns a promise, i.e. every method that is
asynchronous, starts with the word "async". However, all the methods in
AsyncCrypto were an exception to this. I wanted to fix this so that the entire
codebase is consistent. It is only useful that every async method reliabily
starts with the word "async" if it is genuinely the case that every async
method actually starts with the word "async". AsyncCrypto was somewhat of an
exception because, of course, Async is in the title of the class. Nonetheless I
decided to fix this and rename all the methods so that they start with the word
"async". Now that the methods are all consistent, it was redundant to have the
name "Async" in the name "AsyncCrypto". So I have renamed it from AsyncCrypto
to CryptoWorkers, which is slightly more explanatory and has no redundancy.