bip44wallet .toJSON() & .fromJSON()
This is useful for saving a wallet, including its current cache of derived
keys. Note that we will also probably want to add a map of addresses to private
keys in the wallet and/or accounts, and that will also need to be saved to/from
JSON. We will also probably want a way to save the wallet one key at a time,
where the keys are individually stored in the database. That should make saving
changes to the wallet faster. We can worry about that later. For now, the next
step will be to store the entire wallet to the database in one go when it needs
to be saved.