merge blog, docs, ember, landing into main repo
I have found it cumbersome to maintain separate repos for each of our projects.
As before where I merged datt-node and datt-react into a single repo, I'm now
merging our other major projects into the same repo. The other projects are:

- docs: The documentation for Datt.
- blog: The blog content and source.
- ember: The Ember UI, currently unmaintained.
- landing: The landing page, currently hosted at datt.co.

Theoretically, keeping everything in one repo will let us do things like add a
single commit that updates several projects at once, which might be useful for
updating the p2p protocol and its documentation, for instance. This should also
make it easier to manage shared CSS and javascript across different projects.
For instance, all the HTML pages should share the same root CSS. That is not
yet the case, but we can iterate towards that ideal now that everything is in
the same repo.

My motivation for merging everything stems both from my own experience, and
from the fact that Facebook and Google both use one large repo for all
projects. Keeping things unified seems to make things easier to manage.

The commit that merged datt-node and datt-react was called "new initial
commit". This could be called "new initial commit 2".

The root directory of this repo currently is the main application. We may want
to put this into a subdirectory at some point, or perhaps convert "core" and
"react" into their own proper node modules. But not yet. For now, the root
project is the main app, including core and react, and the other projects are
subdirectories which may or may not have their own package.json files.