add tests back into npm package
Having the tests in the npm package is a way to easily test that the npm
package is "really working", i.e. there weren't some weird problems occuring
only during the package creation. The tests are not that big, so it shouldn't
be a big deal that they are being delivered by the npm package even if they
aren't being used most of the time.