proxy insight API for blockchain info
In order to know if you've been paid, you need to be able to query the
blockchain for unspent transaction outputs (UTXOs) that go to your address(es).
Datt also makes use of other features of the blockchain, such as signing the
latest block hash. Insight is an API developed by BitPay that lets us query the
blockchain. It is open-source, and is included in the latest version of
bitcore, which is now a bitcoin full node with a blockchain index. The absolute
simplest thing we can do to get access to the blockchain is simply proxy the
API http requests directly to BitPay's live running instance of Insight. That's
good enough for a prototype. Before actually launching to real users, we'll
have to run our own instance of Insight, and allow the users to run their own
instance and (easily) override the settings of which Insight to use.

When running "gulp serve", you can access the blockchain API at:
- http://localhost:3040/blockchain-api/status?q=getInfo
...where everything after blockchain-api/ is whatever stuff you want to use the
Insight API. See their documentation at:
- https://github.com/bitpay/insight-api

Note that I am not presently forwarding web socket requests, which is necessary
to notice if you've been paid without polling the API. We don't need that for a
prototype, so we can add it later.

Note also that I'm simply assuming we're using mainnet, like I have assumed
elsewhere in the Datt codebase. That is not ideal - we need to find a good way
to distinguish mainnet from testnet in a config option. When you're on testnet,
the Insight URL should be test-insight.bitpay.com.