BlockchainAPI
Create new BlockchainAPI interface to be used to access the blockchain, both
for querying and for sending data (e.g., broadcasting a tx). In the future we
will want to use SPV where possible for most of the same functionality, but for
now, accessing an API is easier.

TODO: The tests are now making real calls to a public blockchain API. We need
to update most tests to use a mocked version of BlockchainAPI since making a
real call is slow.