document environment variables in readme
We're going to want to add more environment variables for config. Next up will
be environment variables that specify whether we are on bitcoin testnet or
mainnet. For now I have documented all the variables that I can remember are
being used somewhere. All of these environment variables are used in the
gulpfile.

We may want to move this info into the docs/ folder at some point, but for now
it seems most convenient to put it in the readme.