display address
We're now displaying the hex version of the address next to each piece of
content. We should really display it in base58 - but that will require
intelligently using the CrypoWorkers module.

Long-term, we won't want to display the address at all. But for now, it is
helpful for debugging purposes and to make sure things are working correctly.