bip44account.fromJSON bug
When getting the list of accounts in a wallet from JSON, I was using the object
properties, which were strings of integers, to set the bip44accounts map.
However, the map was expecting integers, not strings. This meant when going to
retrieve an account, it would not be there, and would be rederived. The end
effective of this is that addresses and private keys were not being cached
properly. Although they were being saved properly, they were not being loaded
properly. This is now fixed. When you load the wallet from the db, the accounts
are set correctly. If you have already derived some addresses, they will still
be there when you load from the db.