testnet support
Bitcoin has two networks, "mainnet" and "testnet". Mainnet is the real bitcoin
network, while testnet is just for testing and has custom rules suitable for
development. It is useful to be able to run Datt in testnet mode so that we
don't have to use real bitcoins when testing it, in case something is broken
and we end up losing money.

You can now run the server such that the client is running in testnet mode with
this command:

FULLNODE_NETWORK=testnet DATT_BLOCKCHAIN_API_URI=https://test-insight.bitpay.com/api/ npm run serve

Note that you must specify both FULLNODE_NETWORK and DATT_BLOCKCHAIN_API_URI,
and the later must end with a "/".

Note also that for right now, testnet and mainnet share the same database. If
you run in mainnet and then run in testnet, you will be using the same
database, which will cause errors. The p2p connections are also the same.
Testnet mode is basically only useful for local testing where you are prepared
to wipe the database after using it.