display deposit address
Note that for right now we are using the 0th address of the 0th account as the
"deposit" address. This is terrible for privacy - we should not be using the
same address for multiple deposits. We should update this so that when you
press the 'receive' button, a new deposit address is created. In any case, this
is good enough for a prototype.