use "ext" and "int" language
It was confusing that you could get "addresses", but that did not include
"change addresses". Rather than stick with the confusing terminology that a
"change address" is not an "address", I've decided to adopt the same language
used by BIP 44: An "external address" is an address you intend to give to
someone else so you can receive money. An "internal address" is an address you
use to receive change in your own addresses - and is not something you give out
to others intended to receive money. For short, "external" -> "ext", and
"internal" -> "int".