add "get all addresses" method
We will need to be able to get the "balance" of an account, which means adding
up all (confirmed or unconfirmed) UTXOs in a wallet. That means we need to have
a list of the addresses in the wallet, including both external (non-change) and
internal (change) addresses. That is what this method is for.