make "receive" button work
The "receive" button now generates a new bitcoin address and displays it. This
is better than the previous behavior where the 0th address was always
displayed. Now, the latest receive address is "kept track of" by the BIP 44
wallet, because only addresses created with "GetNextAddress" are kept track of
and saved to the db. This also means we are not re-using addresses.