GetUTXOsJSON
This gets the raw output of the blockchain API for the UTXOs of a set of
addresses. This is necessary to get the balance of a wallet - once we can get
all UTXOs, we can add up the ones that are confirmed and also optionally the
ones that are unconfirmed. We can also build transactions using these UTXOs.
Right now we only return JSON, but we could also add methods that return
fullnode objects.