poll continuously for bitcoin balance
Updating dattcore and dattreact, we are now polling continuously to get the
balance of the user's bitcoin wallet. Every 10 seconds we use the blockchain
API to query the unspents of all of the users bitcoin addresses. This is not,
long-term, a scalable way to get a bitcoin balance. However, it is good enough
for a prototype, and may work well enough for quite a while in production. We
will see.

The way network connections are "initialized" in dattcore are now broken up
into a separate method, asyncNetworkInitialize(). This is so that the main app
can deliberately be used in "offline" mode if necessary - you can initialize
the database, user, and wallet without establishing network connections.