use more compact "this" reference in asink
The new version of asink matches the "spawn" method in the "official" upcoming
ES7 spec, which allows us to pass in "this" as an argument rather than call
.bind. For one, this slightly shortens the syntax. And two, this place more
nicely with js standard. We can *always* pass in "this", and js standard will
not complain that we are needlessly using .bind.