map from address to keys
A "keys" object contains an xprv, xpub, and address. The BIP 44 Account now
keeps track of a map from address hex strings to keys. This means you can look
up the private key or public key corresponding to an address. The public key
and private keys are on the xpub and xprv objects.