add wallet withdrawal ability to UI
You can now withdraw from the Datt wallet. So the built-in bitcoin wallet is
mostly working now. There are definitely some bugs that will have to be solved
before letting anyone put real money in the wallet, but the basics are all
working. You can create new addresses, deposit money, notice that the money is
deposited, and then withdraw the money to another wallet. We're now very close
to being able to make a payment to a piece of content. Because, of course, if
you can withdraw, you can send to any address, and Datt content already has
bitcoin addresses associated with it.