fix proxy
The proxy I was using, express-request-proxy, was a very sophisticated tool
that seemed to mostly work fine, except that it wasn't relaying POST messages
correctly. This was causing blockchain API POST queries not to work in the
browser. I'm not sure what the problem was, but I got things to work using
http-proxy instead, which is a much simpler tool with fewer features. I also
had to fix the new "local" scripts I created, since they were setting the wrong
environment variable.