copyright datt contributors
The easiest way to deal with the distinction between the company and the
software is to give ownership to the contributors, not the company. My current
strategy is to retire the (almost entirely unused) corporation. Instead, I will
partner with people to make a new corporation that works on Datt, but that has
a different name. Datt will be an open-source project, not a company.

See some thoughts here:
https://ryanxcharlestimes.com/thailand-vs-panama-vs-san-francisco-b4b47d5ca49b