docs: add web torrent to related projects
"web torrent" is an implementation of bit torrent in pure javascript, using Web
RTC for p2p in a browser. It is potentially very useful for Datt. Although it
was already listed in "tools", I have now also added it to "related projects"
as an extra reminder to possibly use this as our p2p layer.

Note that web torrent supports BEP 10, which are extensions to bit torrent that
don't interfere with other bit torrent clients. Conceivably this is the right
way to add all the features we need for Datt.