fullnode@0.11.11
There was a minor breaking change in the new version of fullnode. Basically, in
Datt we were abusing the way to send pubkeys back and forth across workers. The
correct way to do it is to use the new .fromFastBuffer and .toFastBuffer
methods in fullnode, which are specifically designed to serialize an object
into a form suitable for sending to a worker.