new fullnode workers work in node, not browser
The browser stuff is still broken because PeerJS uses "window" which breaks
inside a web worker. The way to fix this is to make it so the web workers don't
load PeerJS, which of course they don't need anyway. I was able to manually fix
it by fixing all of PeerJS's calls to 'window' and replacing with with code
like this:

let win = typeof window === 'undefined' ? self : window