use global fullnode so things work in a browser
I was having a lot of problems using fullnode's new worker interface while
still using "require" everywhere. The basic reason was that fullnode's worker
requires access to the fullnode object inside the worker - so if things weren't
building correctly, that wasn't working. This problem goes away if we just
build fullnode to its own file as a global object and then use that global
object everywhere.