gulp serve
"gulp serve" is used to create a server so you can access the app. It is only
intended for testing/development and not yet ready for deployment. Note that,
unlike before, the new "gulp serve" does not build anything before running. You
need to run "gulp build" to build all files first.