v0.2.3
I just added a whole bunch of commits that change the way the datt bundles are
built, and simultaneously removed the CryptoWorkers interface and replaced it
with the fullnode workers interface. Although the tests pass and the UI seems
to work, some things may be silently broken after all this.