remove hyphen from filenames
Most filenames had a hyphen to distinguish different words, e.g.
db-bip44-wallet.js. I have changed all of these to get rid of the hyphen, e.g.
dbbip44wallet.js. This is purely for consistency with fullnode - the convention
is that filenames do not have hyphenes. There is an exception - the exception
is that files that have different implementations for node and browsers do have
hyphens to distinguish them. The reason for this is to clarify that this is not
a normal class name. For instance, connection-browser-webrtc.js SHOULD have the
hyphens, because it should remplement the same interface as the other
"connection" classes, just that it is for web RTC in a browser.