base datt's p2p message on fullnode's
Rather than implementing essentially the same p2p messages over and over, it's
better to reuse the same messages from fullnode. The only reason why we hadn't
done this before was because of the inability to conveniently asynchronously
create or validate messages. However, with the new fullnode worker interface,
that problem is now fixed. You can create a new message, which involves hashing
the data to compute the checksum, asynchronously. Now the only difference
between a fullnode message and a datt message is that a fullnode message uses
the bitcoin magicnum constant, whereas datt uses a different constant.