replace request with fetch
fetch is smaller than request, is based on promises so it is easier to use, and
is a browser standard ( https://fetch.spec.whatwg.org/ ).

It's worth noting that although request is no longer a direct dependency, it
still occurs in the npm-shrinkwrap file because it is a dependency of pouchdb.