replace q.delay with asyncdelay
The only thing we were using q for was q.delay. Since q is a large package with
lots of other stuff in it, I decided to create my own exceptionally simple
package called "asyncdelay" which does the same thing as q.delay. Now we can
get rid of q entirely. q is a useful package and it may come back later if we
use other packages with callback-based asynchronous interfaces, but for now
this is simpler.