remove p2p and fix other minor issues
While Datt is still intended to have some type of integrated p2p storage
system, the current system is very immature and should be completely replaced
with either IPFS or web torrent. We are moving forward with building a new
product based on Datt, Yours, which does not use Datt's p2p code at all.
Because it adds filesize and because it is going to be replaced anyway, I am
removing it completely. This includes also removing the rendezvous server and
the PeerJS dependency.

I have also had to fix a few other issues that appear to have cropped up
recently:
- The Fullnode Txbuilder now has .toAddress instead of just .to.
- The Insight API seems to break randomly on the scriptSigs it returns, but it
  sometimes parses them into addresses, so I have modified
  GetBlockchainPayerAddresses to accept either case, and also added a TODO to
  get rid of Insight.