fix wallet inconsistency between tabs
The wallet lives in memory and gets saved to an object in the database titled
"bip44wallet". This object has all the addresses and private keys. Although the
addresses and private keys are deterministic, it's important to save them after
deriving them so we know what addresses to watch for transactions, and so we
don't have to do the computational effort of rederiving them later.

There was a problem with the way the wallet was working when different tabs are
used. Different tabs share the same database, but not the same memory. If you
created a new address in one tab, that would not be reflected in the other tab.
Imagine making 2 new addresses in a tab, then making 1 new address in the
second tab. When the second tab saves the wallet file, it overrides the 2
addresses that had been previously created. One gets lost (not permanently
since they are deterministic - but the wallet no longer remembers it exists).

The best solution to this would be to change the architecture of the wallet so
that each address and xpub/xprv pair are saved individually in the database.
However, another solution will suffice for now: checking whether the wallet
object has been changed since the last time we loaded or saved it. If so,
reload it first (because it must have been changed in another tab), and then
create the new address, and then save the wallet file again.

This is good enough for now, but has the consequence that each time you switch
between tabs and then create a new address, the entire wallet has to be loaded
again, which gets laggy if you have many addresses.

I have also updated the serialization of bip32 objects (xpub/xprv) to use
.toFastHex and .fromFastHex, which use only non-blocking cryptography.