index contentauths by time to get recent
We need the ability to get "recent" contentauths. These are contentauths that
we have seen recently - not necessarily that were produced recently. We were
doing this in a very crude way by retrieving the database and storing in
memory, but that is extremely inefficient. The pouchdb-find module is designed
to support indexes and MongoDB-like queries on those indexes. Via pouch-db
find, I have added a "time" field and index, which measures the time in
milliseconds that a piece of content was added to the db, that will now exist
on all contentauths. You can thus retrieve contentauths, limited in number, and
in reverse chronological order, with asyncGetRecentContentAuth, much more
efficiently than before.